backend/requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
psycopg2-binary>=2.9
python-dotenv>=1.0
pgvector>=0.2
sentence-transformers>=3.0
torch==2.5.1
PyPDF2>=3.0
python-docx>=1.1
openai>=1.60
